Base code for Decks working: 7/6/17
Base code for the Deck of Benevolence is up, and it actually works!
(wow)

TODO: Implement actual effects of the Deck of Benevolence, then
implement Deck of Malevolence and Deck of Oddities.

Also come up with a final idea for Trump Card (possibly renamed to Ace
in the Hole) ability.